fix lock test
