+++ src/utils/Client.util.ts
@@ -13,7 +13,10 @@ limitations under the License.
 
 import * as grpc from "@grpc/grpc-js";
 import { KeyValueType } from "../types/KeyValue.type";
-
+import { ConfigurationType } from "../types/KeyConfig.type";
+import { ConfigurationItem } from "../types/configuration/ConfigurationItem";
+import { ConfigurationItem  as ConfigurationItemProto} from "../proto/dapr/proto/common/v1/common_pb";
+import { Map } from "google-protobuf";
 /**
  * Converts a KeyValueType to a grpc.Metadata object.
  * @param metadata key value pair of metadata
@@ -48,4 +51,28 @@ export function createHTTPMetadataQueryParam(metadata: KeyValueType = {}): strin
     // strip the first "&" if it exists
     queryParam = queryParam.substring(1);
     return queryParam;
+}
+
+/**
+ * Converts a Map<string, common_pb.ConfigurationItemProto> to a ConfigurationType object.
+ * @param Map<string, common_pb.ConfigurationItemProto> 
+ * @returns ConfigurationType object
+ */
+ export function createConfigurationType(configDict: Map<string, ConfigurationItemProto>): ConfigurationType {
+    const configMap: { [k: string]: ConfigurationItem } = {};
+
+    configDict.forEach(function(v, k) {
+      const item: ConfigurationItem = {
+        key : k,
+        value : v.getValue(),
+        version : v.getVersion(),         
+        metadata :  v.getMetadataMap().toObject().reduce((result: object, [key, value]) => {
+        // @ts-ignore
+        result[key] = value;
+        return result
+      }, {}),
+      };
+      configMap[k] = item
+  })
+  return configMap
 }
